fix: the mysterious route status is actually the route speed
