+++ packages/core/src/capabilities/Protocols.ts
@@ -2,3 +2,19 @@ export enum Protocols {
 	ZWave = 0,
 	ZWaveLongRange = 1,
 }
+
+export enum ProtocolDataRate {
+	ZWave_9k6 = 0x01,
+	ZWave_40k = 0x02,
+	ZWave_100k = 0x03,
+	LongRange_100k = 0x04,
+}
+
+// Same as ProtocolDataRate, but with the ability to NOT specify a data rate
+export enum RouteProtocolDataRate {
+	Unspecified = 0x00,
+	ZWave_9k6 = 0x01,
+	ZWave_40k = 0x02,
+	ZWave_100k = 0x03,
+	LongRange_100k = 0x04,
+}
