+++ packages/zwave-js/src/Controller.ts
@@ -1,3 +1,4 @@
+export { ProtocolDataRate } from "@zwave-js/core";
 export type {
 	ControllerEvents,
 	HealNodeStatus,
@@ -6,11 +7,6 @@ export type {
 export type { ControllerStatistics } from "./lib/controller/ControllerStatistics";
 export { ZWaveFeature } from "./lib/controller/Features";
 export * from "./lib/controller/Inclusion";
-export {
-	ProtocolDataRate,
-	RSSI,
-	RssiError,
-	TXReport,
-} from "./lib/controller/SendDataShared";
+export { RSSI, RssiError, TXReport } from "./lib/controller/SendDataShared";
 export type { ZWaveLibraryTypes } from "./lib/controller/ZWaveLibraryTypes";
 export { RFRegion } from "./lib/serialapi/misc/SerialAPISetupMessages";
